requirements.txt
pandas
transformers
newspaper3k
python-docx
torch
lxml[html_clean]
python-dotenv
python-telegram-bot==20.7
